create usePostLike hook for reusability
